* master: (3870 commits)
  svn update
  update addon scripts
  file had non utf8 characters.
  quiet compiler warnings for -Wundef
  fix [#27458] Links for Logic Bricks no longer work event->mval wasn't being set in some cases by the event system.
  correct a tooltip misspell
  missed this last commit (was BGE only)
  view3d function naming, no functional changes.
  fix [#27437] Appending armatures with custom bone shapes brings no custom objects into scene.. impossible to edit
  no functional changes, make give_base_to_objects() arguments less confusing.
  Bugfix [#27452] Wrong linking of character groups
  patch [#27449] bpy_script_paths() return path reversed [patch] from Dan Eicher (dna)
  Fix for [#27333] Translation constraint is broken.
  change transform to round at 0.5 for fcurves. this gave noticeable issue when transforming int/fcurve's below zero.
  was using the wrong rna type for GRAPH_OT_select_leftright
  2.5 Particle UI: *In Particle Mode, there was no way to see what p-sys is being edited in the 3D View Tool bar, when having multiple ones. Changed List type to normal (with limit to 3 rows). Request by venomgfx. :)
  error in recent commit.
  edit gp_stroke_convertcoords not to modify the mval passed to it & make some mval args const elsewhere too.
  use consistant arguments to ED_view3d_win_* funcs, a single float vector rather then 2 floats.
  typo in previous commit
  ...

Conflicts:
	extern/bullet2/src/BulletCollision/CollisionDispatch/btCollisionWorld.cpp
	extern/bullet2/src/BulletCollision/CollisionDispatch/btConvexPlaneCollisionAlgorithm.cpp
	extern/bullet2/src/BulletCollision/CollisionDispatch/btGhostObject.h
	extern/bullet2/src/BulletCollision/CollisionShapes/btMaterial.h
	extern/bullet2/src/BulletCollision/CollisionShapes/btMultimaterialTriangleMeshShape.h
	extern/bullet2/src/BulletCollision/NarrowPhaseCollision/btGjkEpaPenetrationDepthSolver.cpp
	extern/bullet2/src/BulletDynamics/ConstraintSolver/btContactConstraint.cpp
	extern/bullet2/src/BulletDynamics/ConstraintSolver/btGeneric6DofConstraint.cpp
	extern/bullet2/src/BulletDynamics/ConstraintSolver/btHingeConstraint.cpp
	extern/bullet2/src/BulletDynamics/ConstraintSolver/btSequentialImpulseConstraintSolver.cpp
	extern/bullet2/src/BulletDynamics/Dynamics/btActionInterface.h
	extern/bullet2/src/BulletDynamics/Dynamics/btDiscreteDynamicsWorld.cpp
	extern/bullet2/src/BulletSoftBody/btSoftBody.cpp
	extern/bullet2/src/BulletSoftBody/btSoftBodyHelpers.cpp
	extern/bullet2/src/LinearMath/btConvexHull.cpp
	extern/bullet2/src/LinearMath/btQuickprof.cpp
	intern/smoke/intern/FLUID_3D.cpp
	source/blender/blenfont/intern/blf.c
	source/blender/blenfont/intern/blf_glyph.c
	source/blender/blenkernel/intern/action.c
	source/blender/blenkernel/intern/depsgraph.c
	source/blender/blenkernel/intern/exotic.c
	source/blender/blenkernel/intern/object.c
	source/blender/blenkernel/intern/particle_system.c
	source/blender/blenkernel/intern/text.c
	source/blender/editors/armature/editarmature.c
	source/blender/editors/armature/poseobject.c
	source/blender/editors/interface/interface_handlers.c
	source/blender/editors/interface/interface_layout.c
	source/blender/editors/interface/interface_utils.c
	source/blender/editors/mesh/editmesh.c
	source/blender/editors/mesh/editmesh_mods.c
	source/blender/editors/mesh/editmesh_tools.c
	source/blender/editors/object/object_edit.c
	source/blender/editors/object/object_transform.c
	source/blender/editors/physics/particle_edit.c
	source/blender/editors/screen/glutil.c
	source/blender/editors/space_logic/logic_window.c
	source/blender/editors/space_outliner/outliner.c
	source/blender/editors/space_sequencer/sequencer_select.c
	source/blender/editors/space_view3d/drawobject.c
	source/blender/editors/transform/transform.c
	source/blender/editors/transform/transform_conversions.c
	source/blender/editors/transform/transform_orientations.c
	source/blender/editors/transform/transform_snap.c
	source/blender/editors/uvedit/uvedit_unwrap_ops.c
	source/blender/gpu/intern/gpu_material.c
	source/blender/modifiers/intern/MOD_array.c
	source/blender/modifiers/intern/MOD_build.c
	source/blender/nodes/intern/CMP_nodes/CMP_lensdist.c
	source/blender/python/intern/bpy_rna.c
	source/blender/render/intern/source/convertblender.c
	source/blender/render/intern/source/rayshade.c
	source/blender/render/intern/source/rendercore.c
	source/blender/render/intern/source/shadeinput.c
	source/blender/render/intern/source/strand.c
	source/blender/windowmanager/intern/wm_draw.c
	source/blender/windowmanager/intern/wm_event_system.c
	source/blender/windowmanager/intern/wm_operators.c
	source/gameengine/BlenderRoutines/BL_KetsjiEmbedStart.cpp

+++ source/blender/imbuf/intern/rotate.c
@@ -1,4 +1,4 @@
-/**
+/*
  *
  * ***** BEGIN GPL LICENSE BLOCK *****
  *
@@ -29,8 +29,15 @@
  * $Id$
  */
 
+/** \file blender/imbuf/intern/rotate.c
+ *  \ingroup imbuf
+ */
+
+
 #include "BLI_blenlib.h"
-#include "BKE_utildefines.h"
+#include "BLI_utildefines.h"
+
+
 #include "MEM_guardedalloc.h"
 
 #include "imbuf.h"
@@ -94,7 +101,7 @@ void IMB_flipy(struct ImBuf * ibuf)
 
 void IMB_flipx(struct ImBuf * ibuf)
 {
-	short x, y, xr, xl, yi;
+	int x, y, xr, xl, yi;
 	float px_f[4];
 	
 	if (ibuf == NULL) return;
